requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
contourpy==1.2.1; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
dataclasses-json==0.6.4; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
django==5.0.4; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
fonttools==4.51.0; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
kiwisolver==1.4.5; python_version >= '3.7'
langchain==0.1.17rc1; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.34; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.47rc1; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.1.4; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.0.1; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.51; python_version < '4.0' and python_full_version >= '3.8.1'
marshmallow==3.21.1; python_version >= '3.8'
matplotlib==3.9.0rc2; python_version >= '3.9'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
networkx==3.3; python_version >= '3.10'
numpy==1.26.4; python_version >= '3.9'
openai==1.23.6; python_full_version >= '3.7.1'
orjson==3.10.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pillow==10.3.0; python_version >= '3.8'
psycopg2==2.9.9; python_version >= '3.7'
pydantic==2.7.1; python_version >= '3.8'
pydantic-core==2.18.2; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pymupdf==1.24.2; python_version >= '3.8'
pymupdfb==1.24.1; python_version >= '3.8'
pyparsing==3.1.2; python_full_version >= '3.6.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.1; python_version >= '3.6'
regex==2024.4.28; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rest-framework-simplejwt==0.0.2; python_version >= '3.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.29; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
tenacity==8.2.3; python_version >= '3.7'
tiktoken==0.6.0; python_version >= '3.8'
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
typing-inspect==0.9.0
urllib3==2.2.1; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
